Unify saml_test.go and saml2_test.go
